move draw methods from utils to draw
